+++ mustang_core/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.3.1
+
+- disable serialization for mustang app config 
+
 ## 3.3.0
 
 - New API and interface to access specific state object